Moved functions to expected files
